requirements.txt
a-cv-imwrite-imread-plus==0.11
a-cv2-easy-resize==0.10
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.0
cachelib==0.10.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
coincurve==18.0.0
colorama==0.4.5
comtypes==1.1.14
contourpy==1.0.7
create-empty-image==0.10
cs50==9.2.5
cv2-collage==0.10
cv2watermark==0.10
cycler==0.11.0
dlib-bin==19.24.1
Flask==2.2.2
Flask-Login==0.6.2
flask-ngrok==0.0.25
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
flatbuffers==23.3.3
fonttools==4.39.3
fpdf==1.7.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
mediapipe==0.9.3.0
MouseInfo==0.1.3
multidict==6.0.4
mysql-connector-python==8.0.32
nltk==3.8.1
numpy==1.24.2
openai==0.27.4
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.4.0
protobuf==3.20.3
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pywhatkit==5.4
pywin32==306
regex==2023.3.23
requests==2.28.2
six==1.16.0
sounddevice==0.4.6
soupsieve==2.4
SpeechRecognition==3.10.0
SQLAlchemy==1.4.46
sqlparse==0.4.4
termcolor==2.3.0
touchtouch==0.11
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.2
wikipedia==1.4.0
yarl==1.8.2
